Move the test methods from JDBCClientTestBase -> JDBCClientTest
