+++ src/main/asciidoc/ruby/index.adoc
@@ -14,7 +14,7 @@ To use this project, add the following dependency to the _dependencies_ section 
 <dependency>
   <groupId>io.vertx</groupId>
   <artifactId>vertx-jdbc-client</artifactId>
-  <version>3.3.0.CR2</version>
+  <version>3.3.0-SNAPSHOT</version>
 </dependency>
 ----
 
@@ -22,7 +22,7 @@ To use this project, add the following dependency to the _dependencies_ section 
 
 [source,groovy,subs="+attributes"]
 ----
-compile 'io.vertx:vertx-jdbc-client:3.3.0.CR2'
+compile 'io.vertx:vertx-jdbc-client:3.3.0-SNAPSHOT'
 ----
 
 == Creating a the client
